+++ packages/mistralai-azure/src/models/components/thinkchunk.ts
@@ -27,6 +27,10 @@ export type ThinkChunk = {
   type: "thinking";
   thinking: Array<TextChunk | ToolReferenceChunk | ReferenceChunk>;
   /**
+   * Signature to replay some reasoning blocks across turns.
+   */
+  signature?: string | null | undefined;
+  /**
    * Whether the thinking chunk is closed or not. Currently only used for prefixing.
    */
   closed?: boolean | undefined;
@@ -56,6 +60,7 @@ export type ThinkChunk$Outbound = {
   thinking: Array<
     TextChunk$Outbound | ToolReferenceChunk$Outbound | ReferenceChunk$Outbound
   >;
+  signature?: string | null | undefined;
   closed?: boolean | undefined;
 };
 
@@ -72,6 +77,7 @@ export const ThinkChunk$outboundSchema: z.ZodType<
       ReferenceChunk$outboundSchema,
     ]),
   ),
+  signature: z.nullable(z.string()).optional(),
   closed: z.boolean().optional(),
 });